Add shell script(run in order)
