+++ Makefile
@@ -137,7 +137,7 @@ tags: $(OBJS) entryother.S _init
 vectors.S: vectors.pl
 	perl vectors.pl > vectors.S
 
-ULIB = ulib.o usys.o printf.o umalloc.o
+ULIB = ulib.o usys.o printf.o umalloc.o shlib.o
 
 _%: %.o $(ULIB)
 	$(LD) $(LDFLAGS) -N -e main -Ttext 0 -o $@ $^
@@ -171,6 +171,7 @@ UPROGS=\
 	_mkdir\
 	_rm\
 	_sh\
+	_msh\
 	_stressfs\
 	_usertests\
 	_wc\
@@ -182,7 +183,7 @@ UPROGS=\
 	
 
 fs.img: mkfs $(UPROGS)
-	./mkfs fs.img README ENV ENV_TEST $(UPROGS)
+	./mkfs fs.img README ENV ENV_TEST STARTCOMMAND $(UPROGS)
 
 -include *.d
 
@@ -249,7 +250,7 @@ qemu-nox-gdb: fs.img xv6.img .gdbinit
 EXTRA=\
 	mkfs.c ulib.c user.h cat.c echo.c forktest.c grep.c kill.c\
 	ln.c ls.c mkdir.c rm.c stressfs.c usertests.c wc.c zombie.c\
-	printf.c umalloc.c setenv.c showenv.c loadenv.c\
+	printf.c umalloc.c setenv.c showenv.c loadenv.c test.c\
 	README dot-bochsrc *.pl toc.* runoff runoff1 runoff.list\
 	.gdbinit.tmpl gdbutil\
