+++ STARTCOMMAND
@@ -1,0 +1,1 @@
+loadenv -s
