+++ fs.h
@@ -50,8 +50,8 @@ struct dinode {
 
 // Directory is a file containing a sequence of dirent structures.
 #define DIRSIZ 14
-#define DIRFLR 16
 
+#define FULLDIR 128
 struct dirent {
   ushort inum;
   char name[DIRSIZ];
@@ -73,7 +73,6 @@ struct envs {
   uint envNum;
 };
 
-char dirname[DIRFLR][DIRSIZ];
-static int dirflr = 0;
-
+static char curDir[FULLDIR] = "~";
+static int dirIndex = 1;
 extern struct envs sysEnv;
