+++ syscall.h
@@ -23,5 +23,6 @@
 #define SYS_date            22
 #define SYS_cwrite          23
 #define SYS_dir             24
+// Add here
 #define SYS_setenv          25 
 #define SYS_getenv          26
